services/ai-worker/requirements.txt
# --- Testing ---
pytest
pytest-mock

# --- Web/API & Cloud ---
requests
boto3
minio

# --- Messaging & Databases ---
pika
pymilvus
redis

# --- Audio & Utilities ---
mutagen==1.47.0
soundfile==0.13.1
librosa==0.11.0
pyrubberband
opencc
snac==1.2.1

# --- High-Level AI Frameworks ---
torch==2.3.1
torchaudio==2.3.1
transformers==4.55.2
accelerate
bitsandbytes==0.43.3
huggingface_hub
google-generativeai
crewai
crewai-tools
openai==1.99.9
langchain-openai

# --- Configuration ---
python-dotenv
pydub

# --- Database Drivers ---
psycopg2-binary

# --- Scheduling & Frameworks ---
APScheduler
Flask
line-bot-sdk